Revert "feat: Add open source red team limits (#5230)"
This reverts commit 27add302cce2d3507ed3480dff9339b2dd238a84.
