+++ drizzle/meta/_journal.json
@@ -134,13 +134,6 @@
       "when": 1755723226650,
       "tag": "0018_fat_vanisher",
       "breakpoints": true
-    },
-    {
-      "idx": 19,
-      "version": "6",
-      "when": 1757026874299,
-      "tag": "0019_new_clint_barton",
-      "breakpoints": true
     }
   ]
 }
