+++ src/app/src/pages/redteam/setup/components/Review.test.tsx
@@ -40,7 +40,6 @@ vi.mock('@promptfoo/redteam/sharedFrontend', () => ({
     plugins: [],
     strategies: [],
   }),
-  getEstimatedProbes: vi.fn().mockReturnValue(10),
 }));
 
 vi.mock('../utils/yamlHelpers', () => ({
