+++ src/app/src/pages/redteam/setup/components/Review.tsx
@@ -2,7 +2,6 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import Code from '@app/components/Code';
 import { useEmailVerification } from '@app/hooks/useEmailVerification';
-import { useProbeLimit } from '@app/hooks/useProbeLimit';
 import { useTelemetry } from '@app/hooks/useTelemetry';
 import { useToast } from '@app/hooks/useToast';
 import YamlEditor from '@app/pages/eval-creator/components/YamlEditor';
@@ -13,7 +12,6 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
 import SaveIcon from '@mui/icons-material/Save';
-import ScienceIcon from '@mui/icons-material/Science';
 import SearchIcon from '@mui/icons-material/Search';
 import StopIcon from '@mui/icons-material/Stop';
 import TuneIcon from '@mui/icons-material/Tune';
@@ -38,17 +36,9 @@ import { useTheme } from '@mui/material/styles';
 import TextField from '@mui/material/TextField';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
-import {
-  isFoundationModelProvider,
-  PROBE_LIMIT_EMAIL,
-  PROBE_LIMIT_URL,
-} from '@promptfoo/constants';
-import {
-  ALLOWED_PROBE_LIMIT_EXCEEDANCE,
-  REDTEAM_DEFAULTS,
-  strategyDisplayNames,
-} from '@promptfoo/redteam/constants';
-import { getEstimatedProbes, getUnifiedConfig } from '@promptfoo/redteam/sharedFrontend';
+import { isFoundationModelProvider } from '@promptfoo/constants';
+import { REDTEAM_DEFAULTS, strategyDisplayNames } from '@promptfoo/redteam/constants';
+import { getUnifiedConfig } from '@promptfoo/redteam/sharedFrontend';
 import { Link } from 'react-router-dom';
 import { useRedTeamConfig } from '../hooks/useRedTeamConfig';
 import { generateOrderedYaml } from '../utils/yamlHelpers';
@@ -56,8 +46,8 @@ import DefaultTestVariables from './DefaultTestVariables';
 import { EmailVerificationDialog } from './EmailVerificationDialog';
 import { LogViewer } from './LogViewer';
 import PageWrapper from './PageWrapper';
-import { RunOptionsContent } from './RunOptions';
-import { getEstimatedDuration } from './strategies/utils';
+import { RunOptions } from './RunOptions';
+import { getEstimatedDuration, getEstimatedProbes } from './strategies/utils';
 import type { RedteamPlugin } from '@promptfoo/redteam/types';
 import type { Job } from '@promptfoo/types';
 
@@ -89,16 +79,8 @@ export default function Review({
   const { config, updateConfig } = useRedTeamConfig();
   const theme = useTheme();
   const { recordEvent } = useTelemetry();
-  const { probeLimit } = useProbeLimit();
   const [isYamlDialogOpen, setIsYamlDialogOpen] = React.useState(false);
   const yamlContent = useMemo(() => generateOrderedYaml(config), [config]);
-  const estimatedProbes = useMemo(() => getEstimatedProbes(config), [config]);
-  const exceedsProbeLimit = useMemo(() => {
-    if (!probeLimit) {
-      return false;
-    }
-    return estimatedProbes > probeLimit.remainingProbes + ALLOWED_PROBE_LIMIT_EXCEEDANCE;
-  }, [estimatedProbes, probeLimit]);
 
   const [isRunning, setIsRunning] = React.useState(false);
   const [logs, setLogs] = React.useState<string[]>([]);
@@ -1082,130 +1064,6 @@ export default function Review({
           </Box>
         </Box>
 
-        {probeLimit && probeLimit.enabled && (
-          <Paper elevation={2} sx={{ p: 3, mb: 3 }}>
-            <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 2 }}>
-              <ScienceIcon color="action" />
-              <Typography variant="h6">Red Team Usage</Typography>
-            </Box>
-
-            <Box sx={{ display: 'flex', alignItems: 'center', gap: 3, mb: 2 }}>
-              <Box>
-                <Typography variant="body2" color="text.secondary">
-                  Used This Month
-                </Typography>
-                <Typography variant="h6">{probeLimit.usedProbes.toLocaleString()}</Typography>
-              </Box>
-              <Divider orientation="vertical" flexItem />
-              <Box>
-                <Typography variant="body2" color="text.secondary">
-                  Remaining
-                </Typography>
-                <Typography
-                  variant="h6"
-                  color={
-                    probeLimit.hasExceeded
-                      ? 'error.main'
-                      : probeLimit.remainingProbes < 5000
-                        ? 'warning.main'
-                        : 'text.primary'
-                  }
-                >
-                  {probeLimit.remainingProbes.toLocaleString()}
-                </Typography>
-              </Box>
-              <Divider orientation="vertical" flexItem />
-
-              <Box>
-                <Typography variant="body2" color="text.secondary">
-                  Est. for This Scan
-                </Typography>
-                <Typography
-                  variant="h6"
-                  color={
-                    estimatedProbes > probeLimit.remainingProbes
-                      ? 'error.main'
-                      : estimatedProbes > probeLimit.remainingProbes * 0.5
-                        ? 'warning.main'
-                        : 'success.main'
-                  }
-                >
-                  {estimatedProbes.toLocaleString()}
-                </Typography>
-              </Box>
-            </Box>
-
-            {estimatedProbes > probeLimit.remainingProbes && (
-              <Alert severity={exceedsProbeLimit ? 'error' : 'warning'} sx={{ mb: 2 }}>
-                <Typography variant="body2">
-                  This scan requires approximately {estimatedProbes.toLocaleString()} probes, but
-                  you only have {probeLimit.remainingProbes.toLocaleString()} remaining.
-                  {exceedsProbeLimit
-                    ? ` The scan cannot be started as it exceeds your limit by more than ${ALLOWED_PROBE_LIMIT_EXCEEDANCE.toLocaleString()} probes.`
-                    : ' The scan may be limited. Consider reducing the number of tests.'}
-                </Typography>
-              </Alert>
-            )}
-
-            <Alert severity="info" sx={{ mt: 2 }}>
-              <Box>
-                <Typography variant="body2" sx={{ mb: 2 }}>
-                  Red teaming uses an LLM to generate probes and grade responses. Our community plan
-                  includes 10,000 probes per month.
-                </Typography>
-
-                <Box
-                  sx={{
-                    p: 2,
-                    borderRadius: 1,
-                    backgroundColor: 'rgba(255, 255, 255, 0.1)',
-                    border: '1px solid rgba(255, 255, 255, 0.2)',
-                  }}
-                >
-                  <Typography variant="subtitle2" sx={{ fontWeight: 'bold', mb: 1 }}>
-                    Need more capacity?
-                  </Typography>
-
-                  <Stack spacing={1}>
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <Typography variant="body2">📧</Typography>
-                      <Typography variant="body2">
-                        Contact our sales team at{' '}
-                        <Link
-                          to={`mailto:${PROBE_LIMIT_EMAIL}`}
-                          style={{ color: 'inherit', textDecoration: 'underline' }}
-                        >
-                          {PROBE_LIMIT_EMAIL}
-                        </Link>
-                      </Typography>
-                    </Box>
-
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <Typography variant="body2">🌐</Typography>
-                      <Typography variant="body2">
-                        Visit{' '}
-                        <Link
-                          to={PROBE_LIMIT_URL}
-                          style={{ color: 'inherit', textDecoration: 'underline' }}
-                        >
-                          {PROBE_LIMIT_URL}
-                        </Link>
-                      </Typography>
-                    </Box>
-
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <Typography variant="body2">💬</Typography>
-                      <Typography variant="body2">
-                        Use the chat in the bottom right corner
-                      </Typography>
-                    </Box>
-                  </Stack>
-                </Box>
-              </Box>
-            </Alert>
-          </Paper>
-        )}
-
         <Paper elevation={2} sx={{ p: 3 }}>
           <Box sx={{ mb: 4 }}>
             <Typography variant="h6" gutterBottom>
@@ -1249,65 +1107,19 @@ export default function Review({
               Run the red team evaluation right here. Simpler but less powerful than the CLI, good
               for tests and small scans:
             </Typography>
-            {probeLimit && probeLimit.hasExceeded && (
-              <Alert severity="error" sx={{ mb: 2 }}>
-                You have exceeded your monthly probe limit ({probeLimit.limit.toLocaleString()}{' '}
-                probes). Please contact {PROBE_LIMIT_EMAIL} to upgrade your account.
-              </Alert>
-            )}
-            {exceedsProbeLimit && !probeLimit?.hasExceeded && (
-              <Alert severity="error" sx={{ mb: 2 }}>
-                {probeLimit?.remainingProbes && probeLimit?.remainingProbes > 0
-                  ? `This scan requires ${estimatedProbes.toLocaleString()} probes but you only have ${probeLimit?.remainingProbes.toLocaleString()} remaining. Please reduce the number of tests or plugins to continue. `
-                  : `You have no probes remaining this month.`}{' '}
-              </Alert>
-            )}
-            {probeLimit &&
-              !probeLimit.hasExceeded &&
-              !exceedsProbeLimit &&
-              probeLimit.remainingProbes < 5000 && (
-                <Alert severity="warning" sx={{ mb: 2 }}>
-                  You have {probeLimit.remainingProbes.toLocaleString()} probes remaining this month
-                  ({Math.round((probeLimit.remainingProbes / probeLimit.limit) * 100)}% of your
-                  monthly limit).
-                </Alert>
-              )}
             <Box sx={{ mb: 2 }}>
               <Box sx={{ display: 'flex', gap: 2, alignItems: 'center' }}>
-                <Tooltip
-                  title={
-                    probeLimit?.hasExceeded
-                      ? 'Monthly probe limit exceeded. Please contact inquiries@promptfoo.dev to upgrade.'
-                      : exceedsProbeLimit
-                        ? `This scan requires ${estimatedProbes.toLocaleString()} probes but you only have ${probeLimit?.remainingProbes.toLocaleString()} remaining. Please reduce the number of tests.`
-                        : ''
+                <Button
+                  variant="contained"
+                  color="primary"
+                  onClick={handleRunWithSettings}
+                  disabled={isRunning}
+                  startIcon={
+                    isRunning ? <CircularProgress size={20} color="inherit" /> : <PlayArrowIcon />
                   }
-                  placement="top"
                 >
-                  <span>
-                    <Button
-                      variant="contained"
-                      color="primary"
-                      onClick={handleRunWithSettings}
-                      disabled={isRunning || probeLimit?.hasExceeded || exceedsProbeLimit}
-                      startIcon={
-                        isRunning ? (
-                          <CircularProgress size={20} color="inherit" />
-                        ) : (
-                          <PlayArrowIcon />
-                        )
-                      }
-                    >
-                      {isRunning
-                        ? 'Running...'
-                        : probeLimit?.hasExceeded
-                          ? 'Limit Exceeded'
-                          : exceedsProbeLimit
-                            ? 'Exceeds Limit'
-                            : 'Run Now'}
-                    </Button>
-                  </span>
-                </Tooltip>
+                  {isRunning ? 'Running...' : 'Run Now'}
+                </Button>
                 {isRunning && (
                   <Button
                     variant="contained"
