+++ src/app/src/pages/redteam/setup/components/Strategies.tsx
@@ -17,7 +17,6 @@ import {
   strategyDescriptions,
   strategyDisplayNames,
 } from '@promptfoo/redteam/constants';
-import { getEstimatedProbes } from '@promptfoo/redteam/sharedFrontend';
 import { Link as RouterLink } from 'react-router-dom';
 import { useRedTeamConfig } from '../hooks/useRedTeamConfig';
 import PageWrapper from './PageWrapper';
@@ -27,7 +26,7 @@ import { RecommendedOptions } from './strategies/RecommendedOptions';
 import { StrategySection } from './strategies/StrategySection';
 import { SystemConfiguration } from './strategies/SystemConfiguration';
 import { type PresetId, STRATEGY_PRESETS, type StrategyPreset } from './strategies/types';
-import { getStrategyId } from './strategies/utils';
+import { getEstimatedProbes, getStrategyId } from './strategies/utils';
 import type { RedteamStrategyObject } from '@promptfoo/redteam/types';
 
 import type { ConfigDialogState, StrategyCardData } from './strategies/types';
