+++ src/redteam/commands/run.ts
@@ -8,7 +8,6 @@ import logger from '../../logger';
 import telemetry from '../../telemetry';
 import { setupEnv } from '../../util';
 import { getConfigFromCloud } from '../../util/cloud';
-import { checkProbeLimit } from '../../util/redteamProbeLimit';
 import { doRedteamRun } from '../shared';
 import { poisonCommand } from './poison';
 import type { Command } from 'commander';
@@ -87,14 +86,6 @@ export function redteamRunCommand(program: Command) {
       }
 
       try {
-        // Check monthly probe limit before running redteam
-        const probeCheckResult = await checkProbeLimit();
-
-        if (!probeCheckResult.canProceed) {
-          process.exitCode = probeCheckResult.exitCode;
-          return;
-        }
-
         if (opts.remote) {
           cliState.remote = true;
         }
