+++ src/redteam/constants.ts
@@ -2,6 +2,3 @@ export * from './constants/frameworks';
 export * from './constants/metadata';
 export * from './constants/plugins';
 export * from './constants/strategies';
-
-export const MONTHLY_PROBE_LIMIT = 10000;
-export const ALLOWED_PROBE_LIMIT_EXCEEDANCE = 100;
