+++ src/server/apiSchemas.ts
@@ -31,15 +31,6 @@ export const ApiSchemas = {
         message: z.string().optional(),
       }),
     },
-    ProbeLimit: {
-      Response: z.object({
-        hasExceeded: z.boolean(),
-        usedProbes: z.number(),
-        remainingProbes: z.number(),
-        limit: z.number(),
-        enabled: z.boolean(),
-      }),
-    },
   },
   Eval: {
     UpdateAuthor: {
